feat: update analyzer plot
